backend/requirements.txt
anthropic==0.18.1
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pandas==2.2.1
numpy==1.26.4 